Moved tests under project-specific folder for cleaner structure.
